Fix address_of for 64-bit byte indexing.
Former-commit-id: a77c984cc3268ac30e8ed60dc1be23249c772d5e

+++ src/Image.h
@@ -157,7 +157,8 @@ public:
     /** Get the address of a particular pixel. */
     void *address_of(int x, int y = 0, int z = 0, int w = 0) const {
         uint8_t *ptr = (uint8_t *)origin;
-        size_t offset = x*stride_0 + y*stride_1 + z*stride_2 + w*stride_3;
+        uint64_t offset = (int64_t)x*stride_0 + (int64_t)y*stride_1 +
+            (int64_t)z*stride_2 + (int64_t)w*stride_3;
         return (void *)(ptr + offset * elem_size);
     }
 };
